+++ packages/gatsby-source-drupal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@4.6.0-next.3...gatsby-source-drupal@4.7.0-next.0) (2021-05-21)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+# [4.6.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@4.6.0-next.2...gatsby-source-drupal@4.6.0-next.3) (2021-05-19)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
 # [4.6.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@4.6.0-next.1...gatsby-source-drupal@4.6.0-next.2) (2021-05-19)
 
 ### Features